+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### New features
 
+ - Make `\WizaplaceFrontBundle\Twig\AppExtension::formatPrice` use the `currency` translation key instead of `€`.
  - Upgrade SDK to v1.15.1
 
 ### Bugfixes